Divided VoteService from VotingService, added new methods to VotingService
