+++ BLL/Interfaces/IVotingService.cs
@@ -9,10 +9,8 @@ namespace BLL.Interfaces
 {
     interface IVotingService : ICRUD<VotingModel>
     {
-        Task AddVoteAsync(VoteModel model);
+        IEnumerable<VotingModel> GetFilteredAndSortedForUser(string userId);
 
-        IEnumerable<VoteModel> GetAllVotes();
-
-        Task<VoteModel> GetVoteByIdAsync(int id);
+        IEnumerable<VotingModel> GetFilteredAndSortedForAdmin();
     }
 }
